requirements.txt
cffi==2.0.0
charset-normalizer==3.4.4
coverage==7.13.0
cryptography==46.0.3
Deprecated==1.3.1
fonttools==4.61.1
iniconfig==2.3.0
lxml==6.0.2
lxml_html_clean==0.4.3
# Editable Git install with no remote (mcps==0.1)
-e /home/michel/Desktop/mcps
packaging==25.0
pikepdf==10.1.0
pillow==12.0.0
pluggy==1.6.0
pycparser==2.23
Pygments==2.19.2
pypdf==6.5.0
PyPDFForm==3.8.2
pyperclip==1.11.0
pytest==9.0.2
pytest-cov==7.0.0
reportlab==4.4.7
wrapt==2.0.1
PyYAML==6.0
